Add state tracking for event recording
Following scenarios can lead to wrong event recording:

- a function may be called by multiple functions
- buffer overrun (then we miss certain sub-event of the event)
- some BPF progs are not attached yet (happens at early stage of profiling)

In light of this, we introduce state tracking for event recording, with this,
we simply discard the whole event if one of the sub-event is lost, or if the
sub event is triggered by caller we are not interested in, we don't record the
event. The same is true at the early stage, where some BPF progs may not loaded
yet.